更改启动磁盘位置从E盘到D盘，以及定时清除缓存

+++ src/main/java/com/entity/analysis_voice.java
@@ -11,7 +11,7 @@ public class analysis_voice {
 
     private Integer numberOfPauses;
 
-    private Integer rateOfSpeech;
+    private Float rateOfSpeech;
 
     private Float articulationRate;
 
@@ -77,11 +77,11 @@ public class analysis_voice {
         this.numberOfPauses = numberOfPauses;
     }
 
-    public Integer getRateOfSpeech() {
+    public Float getRateOfSpeech() {
         return rateOfSpeech;
     }
 
-    public void setRateOfSpeech(Integer rateOfSpeech) {
+    public void setRateOfSpeech(Float rateOfSpeech) {
         this.rateOfSpeech = rateOfSpeech;
     }
